Checkin note cleanup
