+++ server/src/main/java/com/objectcomputing/checkins/services/checkin_notes/CheckinNoteRepository.java
@@ -1,7 +1,4 @@
-package com.objectcomputing.checkins.services.checkinnotes;
-
-import java.util.List;
-import java.util.UUID;
+package com.objectcomputing.checkins.services.checkin_notes;
 
 import io.micronaut.data.jdbc.annotation.JdbcRepository;
 import io.micronaut.data.model.query.builder.sql.Dialect;
@@ -10,6 +7,8 @@ import io.micronaut.data.repository.CrudRepository;
 import javax.annotation.Nonnull;
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
+import java.util.List;
+import java.util.UUID;
 
 
 @JdbcRepository(dialect = Dialect.POSTGRES)
