+++ server/src/test/java/com/objectcomputing/checkins/services/checkin_notes/CheckinNotesBadArgExceptionTest.java
@@ -1,4 +1,5 @@
-package com.objectcomputing.checkins.services.checkinnotes;
+package com.objectcomputing.checkins.services.checkin_notes;
+
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
